Implementation of Binary Search Tree (BST)
This commit contains a Ruby implementation of a BST. The class can construct a BST from a given input array. It can search the BST using BFS or DFS. Some statements are printed for debugging and understanding what is going on during writing of the code. Could be optimized for 'Rubyness'...